Added check for blank profile name
Added check for blank profile name. This is, of course, in preperation
for the next big planned feature for Mr. BallGuy: Profiles! There are
still some debugging things left in the game, but other than that
everything should work fine as for right now. Please report any issues.

Oh, and some minor bug fixes as well.